refactor controls hook + views & add TextureView
