+++ src/components/SceneView.tsx
@@ -20,7 +20,7 @@ import {
   selectUneditedTextureUrls
 } from '@/store/selectors';
 import { setObjectKey, useAppDispatch, useAppSelector } from '@/store';
-import { useSceneKeyboardControls } from '@/hooks';
+import { useObjectNavControls } from '@/hooks';
 import ViewOptionsContext from '@/contexts/ViewOptionsContext';
 import { useTheme } from '@mui/material';
 import { SceneContextSetup } from '@/contexts/SceneContext';
@@ -39,7 +39,7 @@ import {
   Vector2
 } from 'three';
 import { objectUrlToBuffer } from '@/utils/data';
-import RenderedPolygon from './RenderedPolygon';
+import RenderedPolygon from './scene/RenderedPolygon';
 
 THREE.ColorManagement.enabled = true;
 
@@ -83,8 +83,8 @@ async function createTextureFromObjectUrl(
 
 const axesHelper = <axesHelper args={[50]} />;
 
-export default function SceneCanvas() {
-  useSceneKeyboardControls();
+export default function SceneView() {
+  useObjectNavControls();
   const [textureMap, setTextureMap] = useState<Map<string, DataTexture>>();
   const canvasRef = useRef() as MutableRefObject<HTMLCanvasElement>;
   const viewOptions = useContext(ViewOptionsContext);
