+++ src/components/TextureView.tsx
@@ -1,0 +1,9 @@
+import { useObjectNavControls } from '@/hooks';
+
+export default function TextureView() {
+    useObjectNavControls();
+    
+    return (
+        <div>Texture View</div>
+    );
+}
