+++ src/hooks/index.ts
@@ -1,4 +1,4 @@
-export { default as useSceneKeyboardControls } from './useSceneKeyboardControls';
+export { default as useObjectNavControls } from './useObjectNavControls';
 export { default as useModelSelectionExport } from './useModelSelectionExport';
 export { default as useHeldRepetitionTimer } from './useHeldRepetitionTimer';
 export { default as useDebouncedEffect } from './useDebouncedEffect';
